Data.Word.N

Description

Fixed size bit vectors using type-level naturals.

The W newtype

data W n

Type representing a sequence of n bits, or a non-negative integer smaller than 2^n.

Operations

(>+<) :: forall m n o. Triplet m n o => W m -> W n -> W o infixr 6

Appends two W's, treating the first's bits as more significant.

Example usage:

   import Network.Socket

   fromHostAddress6 :: HostAddress6 -> W 128
   fromHostAddress6 (a, b, c, d) = f a >+< f b >+< f c >+< f d
     where
       f = fromIntegral :: Word32 -> W 32

split :: forall m n o. Triplet m n o => W o -> (W m, W n)

The inverse of >+<

   forall a b. split (a >+< b) == (a, b)

Example usage:

   import Network.Socket

   toHostAddress6 :: W 128 -> HostAddress6
   toHostAddress6 w =  (f a, f b, f c, f d)
     where
       f = fromIntegral :: W 32 -> Word32
       (a, x) = split w
       (b, y) = split x
       (c, d) = split y

Conversion Classes

class ToW n a | a -> n where

Provides a way to convert from fixed-bit types to the W n type.

Methods

toW :: a -> W n

class FromW n a | a -> n where

Provides a way to convert from the W n type to fixed-bit types.

Methods

fromW :: W n -> a

type family AccSums list :: Constraint

Convenience constraint synonym. Tuple instances of ToW and FromW would be a lot uglier without it.

Equations

AccSums `[]` = ()

AccSums (n : ns) = (Triplet n (Sum ns) (Sum (n : ns)), AccSums ns)

Convenience

type BothKnown n = (KnownNat n, KnownNat (2 ^ n))

Synonym for readability

type Triplet m n o = (BothKnown m, BothKnown n, BothKnown o, (m + n) ~ o)

Synonym for readability