Data.Word.N.Church

Description

Provides a church-encoding means of inspecting and constructing bit-vectors. This module is currenty a work in progress.

Core types

data family Fun list result :: *

Fn is not injective, but Fun is, which is necessary for type decidability. However, this abstraction is not easily optimized away by GHC, so this will soon change.

class Church list where

Class for nonempty type-level lists of Nat's. This is the core of the heterogeneous-church-encoded-vector-like interface.

Methods

construct :: Fun list (Wof list)

Given components, return their concatenation.

inspect :: Fun list result -> Wof list -> result

Operate on a bit-vector component-wise, where the size of its components are determined by list.

Convenience

type Fn list result = Foldr (->) result (Map W list)

Type synonym for readability. Constructs a the type of functions from the pattern of W's specified by list to result.

type Wof list = W (Sum list)

Type synonym for readability

type ListSum n ns = Triplet n (Sum ns) (n + Sum ns)

Constraint synonym for readablility