Data.Word.N.Conversion

Description

Classes for converting to and from the W n type. ToW and FromW interpret the leftmost members of tuples as the most significant.

Documentation

class ToW n a | a -> n where

Provides a way to convert from fixed-bit types to the W n type.

Methods

toW :: a -> W n

class FromW n a | a -> n where

Provides a way to convert from the W n type to fixed-bit types.

Methods

fromW :: W n -> a

type family AccSums list :: Constraint

Convenience constraint synonym. Tuple instances of ToW and FromW would be a lot uglier without it.

Equations

AccSums `[]` = ()

AccSums (n : ns) = (Triplet n (Sum ns) (Sum (n : ns)), AccSums ns)