Data.Word.N.Util

Description

Useful functions for working with W n's.

Documentation

slice :: forall m n o no mno. (Triplet n o no, Triplet m no mno) => Proxy o -> W mno -> W n

Focused Operations

class (KnownNat d, KnownNat n) => d :|: n where

Reads "d divides n", and allows for generic construction and inpsection.

Methods

assemble :: forall f. Applicative f => (forall a b c. (Triplet a b c, Triplet b a c) => W a -> W b -> W c) -> f (W d) -> f (W n)

Construct an applicative action that results in a W n from one that results in a W d, where d | n. See assembleR and assembleL for examples.

disassemble :: forall m. Monoid m => (W d -> m) -> W n -> m

Construct a means of pushing a W n into a monoid from a means of pushing a W d into the monoid, where d | n. See disassembleR and disassembleL for examples.

assembleL :: (Applicative f, d :|: n) => f (W d) -> f (W n)

Same as assembleR, but treats the first results as less significant.

assembleL = 'assemble' (flip ('>+<'))

Here's the example above, modified to parse in little-endian byte order:

   import Data.Attoparsec.ByteString
   import Data.Word

   anyWord128LE :: Parser (W 128)
   anyWord128LE = assembleL $ fmap (fromIntegral :: Word8 -> W 8) anyWord8

assembleR :: (Applicative f, d :|: n) => f (W d) -> f (W n)

Transforms an applicative action that results in a W d to on that results in a W n, provided that d|n (hence the :|: constraint), treating the first results as more significant.

assembleR = 'assemble' ('>+<')

Example using attoparsec to parse a big-endian unsigned 128-bit integer:

   import Data.Attoparsec.ByteString
   import Data.Word

   anyWord128BE :: Parser (W 128)
   anyWord128BE = assembleR $ fmap (fromIntegral :: Word8 -> W 8) anyWord8

disassembleL :: (Monoid m, d :|: n) => (W d -> m) -> W n -> m

Same as disassembleL, but combines less significant chunks first.

disassembleL f = getDual . disassemble (Dual . f)

disassembleR's example adjusted to build in little-endian byte order:

   import Data.ByteString.Builder

   word128LE :: W 128 -> Builder
   word128LE = disassembleL (word8 . (fromIntegral :: W 8 -> Word8))

disassembleR :: (Monoid m, d :|: n) => (W d -> m) -> W n -> m

Breaks a W n into its constituent d-sized chunks, and combines them according to the provided monoid. More significant chunks are combined first.

disassembleR = disassemble

Example using a bytestrings

   import Data.ByteString.Builder

   word128BE :: W 128 -> Builder
   word128BE = disassembleR (word8 . (fromIntegral :: W 8 -> Word8))

Displaying

showBits :: 1 :|: n => W n -> String

Displays a bitvector's bits in big-endian order.

showHex :: 4 :|: n => W n -> String

Displays a bitvector's bits in big-endian order, in groups of four as hexadecimal digits.