Data.Word.N.Conversion

Description

Classes for converting to and from the W n type.

Documentation

class ToW n a | a -> n where

Provides a way to convert from fixed-bit types to the W n type.

Methods

toW :: a -> W n

class FromW n a | a -> n where

Provides a way to convert from the W n type to fixed-bit types.

Methods

fromW :: W n -> a

newtype BigEndian a

Wrapper for tuples. ToW and FromW interpret the leftmost members of such tuples as the most significant.

Constructors

BigEndian

Fields

fromBigEndian :: a

newtype LittleEndian a

Wrapper for tuples. ToW and FromW interpret the rightmost members of such tuples as the most significant.

Constructors

LittleEndian

Fields

fromLittleEndian :: a